+++ beakpeek/lib/bird.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: avoid_print, library_private_types_in_public_api
 
 import 'dart:convert';
+import 'package:beakpeek/Module/bird.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter/material.dart';
 
@@ -200,37 +201,6 @@ Future<List<Bird>> fetchBirds(String pentadId, http.Client client) async {
     print('Error fetching birds: $error');
     throw Exception('Failed to load birds: $error');
   }
-}
-
-class Bird {
-  Bird({
-    required this.pentad,
-    required this.spp,
-    required this.commonGroup,
-    required this.commonSpecies,
-    required this.genus,
-    required this.species,
-    required this.reportingRate,
-  });
-
-  factory Bird.fromJson(Map<String, dynamic> json) {
-    return Bird(
-      pentad: json['pentad'],
-      spp: json['spp'],
-      commonGroup: json['common_group'],
-      commonSpecies: json['common_species'],
-      genus: json['genus'],
-      species: json['species'],
-      reportingRate: json['reportingRate'],
-    );
-  }
-  final String pentad;
-  final int spp;
-  final String commonGroup;
-  final String commonSpecies;
-  final String genus;
-  final String species;
-  final double reportingRate;
 }
 
 class BirdList extends StatelessWidget {
